% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atomNames.R
\docType{data}
\name{atomNames}
\alias{atomNames}
\title{Atom names/types}
\format{A data frame with the following columns.
  \describe{
    \item{\code{atmname}}{a character vector. Atom names/types.}
    \item{\code{symb}}{a character vector. Atomic symbols.}
}}
\usage{
atomNames
}
\description{
This data set provides various atom names/types and there associated atomic
symbols which are used for atom recognition.
}
\examples{
data(atomNames)
atomNames

}
\keyword{datasets}

